refactor: extract useCharacterForm utilities to reduce complexity
- Create character-form-utils.ts with validation and transformation logic
- Extract categorizeError, validateFormData, and isFormValid functions
- Move all form data types and interfaces to utilities
- Simplify useCharacterForm hook by using extracted utilities
- Reduce complexity from 11 to ~3 by breaking down complex validation logic
- Maintain 100% test coverage for refactored hook

Relates to #12
